refactor: 자동 로그인 위치 App으로 이동
